requirements.txt
streamlit
langchain
langchain-huggingface
langchain-core
python-dotenv
huggingface-hub
